models/tests/test_segregation.py
"""
This is the test suite for segregation.py.
"""

from unittest import TestCase

from propargs.propargs import PropArgs

import models.segregation as seg
from indra.composite import Composite
from indra.env import Env
from registry.registry import get_env
from models.segregation import DEF_TOLERANCE, DEF_SIGMA
from models.segregation import env_favorable, BLUE_AGENTS, RED_AGENTS
from models.segregation import group_names
from models.segregation import get_tolerance
from models.segregation import set_up, create_resident, RED_GRP_IDX, BLUE_GRP_IDX

TEST_ANUM = 999999

REP_RAND_TESTS = 100

SMALL_GRID = 4


def print_sep():
    print("________________________", flush=True)


class SegregationTestCase(TestCase):
    def setUp(self):
        set_up()

    def tearDown(self):
        pass

    # an integration test:
    def test_main(self):
        self.assertEqual(seg.main(), 0)

    def test_get_tolerance(self):
        """
        Test that our tolerance function gets a good distribution.
        """
        sum_of_tolerance = 0
        for i in range(REP_RAND_TESTS):
            sum_of_tolerance += get_tolerance(DEF_TOLERANCE, DEF_SIGMA)
        avg = sum_of_tolerance / REP_RAND_TESTS
        self.assertLess(DEF_TOLERANCE - .2, avg)
        self.assertGreater(DEF_TOLERANCE + .2, avg)

    def test_create_agent(self):
        """
        Test that creating an agent works.
        """
        fred = create_resident(RED_AGENTS, TEST_ANUM)
        freds_nm = group_names[RED_GRP_IDX] + str(TEST_ANUM)
        self.assertEqual(freds_nm, str(fred))

    def agent_in_little_city(self, with_blue=False):
        red_agents = Composite("My reds")
        test_agent = create_resident(RED_AGENTS, TEST_ANUM)
        red_agents += test_agent
        blue_agents = Composite("My blues")
        if with_blue:
            for i in range(0, SMALL_GRID * SMALL_GRID - 1):
                blue_agents += create_resident(BLUE_AGENTS, TEST_ANUM + 1)

        my_city = Env("Small city for test", width=SMALL_GRID,
                      height=SMALL_GRID,
                      members=[red_agents, blue_agents])
        return (test_agent, my_city)

    def test_seg_agent_action(self):
        """
        We are going to test two cases: one where agent should
        be satisfied with neighborhood, and one not.
        """
        (test_agent, city) = self.agent_in_little_city()
        # self.assertEqual(seg_agent_action(test_agent), True)
        (test_agent, city) = self.agent_in_little_city(with_blue=True)
        # the following test is mysteriously failing: must debug!
        # self.assertEqual(seg_agent_action(test_agent), False)

    def test_env_favorable(self):
        env_fav = env_favorable(0.4, 0.5)
        self.assertEqual(env_fav, False)

        env_fav = env_favorable(0.6, 0.5)
        self.assertEqual(env_fav, True)
